personal-color-ai-analysis-main/personal-color-ai-analysis-main/requirements.txt
streamlit
torch
torchvision
Pillow
pandas
opencv-python
scikit-image
